+++ sulong/mx.sulong/suite.py
@@ -440,6 +440,33 @@ suite = {
       "testProject" : True,
       "defaultBuild" : False,
     },
+    "com.oracle.truffle.llvm.tests.bitcodeformat": {
+      "subDir": "tests",
+      "native": True,
+      "vpath": True,
+      "results": [
+        "bitcodeformat/hello-linux-emit-llvm.bc",
+        "bitcodeformat/hello-linux-compile-fembed-bitcode.o",
+        "bitcodeformat/hello-linux-link-fembed-bitcode",
+        "bitcodeformat/hello-linux-link-fembed-bitcode.so",
+        "bitcodeformat/hello-darwin-emit-llvm.bc",
+        "bitcodeformat/hello-darwin-compile-fembed-bitcode.o",
+        "bitcodeformat/hello-darwin-link-fembed-bitcode",
+        "bitcodeformat/hello-darwin-link-fembed-bitcode.dylib",
+      ],
+      "buildEnv": {
+        "SUITE_CFLAGS": "<clangImplicitArgs>",
+        "SUITE_CPPFLAGS": "-I<sulong_include> -I<path:SULONG_LIBS>",
+      },
+      "dependencies": [
+        "SULONG_TEST",
+      ],
+      "buildDependencies": [
+        "SULONG_LIBS",
+      ],
+      "testProject": True,
+      "defaultBuild": False,
+    },
     "gcc_c" : {
       "subDir" : "tests/gcc",
       "class" : "ExternalTestSuite",
@@ -704,6 +731,7 @@ suite = {
       "platformDependent" : True,
       "output" : "mxbuild/<os>-<arch>/sulong-test-suites",
       "dependencies" : [
+        "com.oracle.truffle.llvm.tests.bitcodeformat",
         "com.oracle.truffle.llvm.tests.debug",
         "com.oracle.truffle.llvm.tests.irdebug",
         "com.oracle.truffle.llvm.tests.interop",