+++ sulong/projects/com.oracle.truffle.llvm/src/com/oracle/truffle/llvm/Runner.java
@@ -850,7 +850,7 @@ public final class Runner {
             parserResults.add(parserResult);
             return parserResult;
         } else if (!library.isNative()) {
-            throw new LLVMParserException("The file '" + source.getName() + "' is not a bitcode file nor an ELF File with an .llvmbc section.");
+            throw new LLVMParserException("The file '" + source.getName() + "' is not a bitcode file nor an ELF or Mach-O object file with an embedded bitcode section.");
         } else {
             return null;
         }